services/session_monitor/requirements.txt
requests>=2.31.0
openai>=1.14.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
azure-storage-blob>=12.19.0
